pyproject.toml
[tool.poetry]
name = "buena-fiesta"
version = "0.0.1"
description = "New generation of social network for helping international students - used by sections of Erasmus Student Network."
authors = ["ESN VUT BRNO"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"

Pillow = "^9.0.1"
bcrypt = "^3.2.0"
django-allauth = "^0.47.0"
# support for dj4.0
django-allauth-cas = { git = "https://github.com/thejoeejoee/django-allauth-cas", branch = "master" }
django-click = "^2.3.0"
django-configurations = { extras = ["cache", "database", "email", "search"], version = "^2.4.1" }
django-countries = "^7.2.1"
django-extensions = "^3.1.5"
django-filter = "^21.1"
django-htmx = "^1.14.0"
django-lifecycle = "^1.0.0"
django-loginas = "^0.3.10"
django-model-path-converter = "^0.1.0"
django-polymorphic = "^3.1.0"
django-tables2 = "^2.5.3"
django-webpack-loader = "^1.6.0"
elasticsearch = "^7.17.0"
factory-boy = "^3.2.1"
mysqlclient = "^2.1.0"
psycopg2-binary = "^2.9.3"
python-magic = "^0.4.25"
tablib = { extras = ["xlsx"], version = "^3.2.1" }
faker_education = "^1.2"
django-select2 = "^7.10.0"
django-debug-toolbar = "^3.5.0"
django-watchfiles = "^0.0.1"
django-autocomplete-light = "^3.9.4"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "^6.3.0" }
requests = "^2.28.2"
beautifulsoup4 = "^4.12.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
pydotplus = "^2.0.2"
faker_education = "^1.2"
pydot = "^1.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = 'black'
lines_between_types = 0
combine_as_imports = true
line_length = 120
known_first_party = ["fiesta", "apps"]
known_local_folder = ["apps"]
known_third_party = ["django"]
skip = ["migrations"]

[tool.black]
line-length = 120
target-version = ['py311']
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | migrations
  | build
  | dist
)/
'''

[tool.ruff]
# see https://beta.ruff.rs/docs/rules/
extend-select = ["UP", "DJ", "PIE", "INT", "PTH", "SIM", "RET", "G", "DTZ", "B", "I002"]
line-length = 120
target-version = "py311"
ignore = ["E731"]

exclude = [
    "migrations",
]

[tool.ruff.isort]
required-imports = ["from __future__ import annotations"]


[tool.vulture]
paths = ["fiesta"]
exclude = ["settings"]
min_confidence = 80
make_whitelist = true
